requirements.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-cors-headers==3.0.2
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-taggit==1.1.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
oauthlib==3.0.1
openapi-codec==1.3.2
Pillow==6.0.0
psycopg2==2.8.3
PyJWT==1.7.1
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
